Implements functionality requested in issue #71.
In short, this commit adds a "clone_repo_to" option to allow a user
to provide a directory in which to place clone a remote repo. This
enables the user to access the files within that directory that
aren't altered by a specific commit.

A corresponding test has also been added.