Resolve mypy warnings in RepositoryMining constructor
Public  API constructor arguments `only_modifications_with_file_types` and
`only_commits` take an List as input, but are transformed in the implementation
into a set.

MyPy did not like that the same variable was  used  for  the set and the list.
Introduced a separate one to document the conversion more explicitly.

The alternative option would be to change the public API into expecting
a set rather than a List[str].