Ignore changes in unsupported languages in DMM
Commits can contain modificiations in languages not supported by Lizard.
Such changes should be ignored when computing the delta maintainability metrics.

We add a predicate to Modifications checking whether the programming language used
is supported (this is an extension of the public API), implemented by a call to Lizard.

If none of the modifications of a commit is written in a supported language,
the dmm value returned becomes None (this makes the return type Optional[Float]
instead of just Float).