Add test repository for mailmap feature
Together with the feature implementation, I provide a set of tests. They rely
on an example repository, which I adapted from
[here](https://github.com/ContentMine/mailmap). I needed a small repository
with a `.mailmap` file and few commits in which some are made by users with
different name/email values. I do not have the time to craft one from scratch.
Therefore, I hope it is okay to include that repository in the test
repositories. I believe there is no issue, since the tests as well as the
commit message state the source of the data explicitly.